Fix int overflow when casting period_ms*1000 to chrono::microseconds
period_ms is a float read from the topology json. For topologies with
extremely long idle periods (e.g. 1h period_ms = 3600000, microseconds
= 3.6e9) the int cast overflowed signed-int32, producing nonsense
periods. Use int64_t to match the underlying chrono::microseconds rep.

Applies to the publisher, client, and action-client periodic factories.

+++ performance_test_factory/src/factory.cpp
@@ -549,7 +549,7 @@ void TemplateFactory::add_periodic_publisher_from_json(
     std::cout << "Error! Publishers must set period_ms or freq_hz in json file" << std::endl;
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   size_t msg_size = 0;
   if (pub_json.find("msg_size") != pub_json.end()) {
@@ -612,7 +612,7 @@ void TemplateFactory::add_periodic_client_from_json(
     std::cout << "Error! Clients must set period_ms or freq_hz in json file" << std::endl;
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   rclcpp::QoS custom_qos_profile = get_qos_from_json(client_json);
 
@@ -641,7 +641,7 @@ void TemplateFactory::add_periodic_action_client_from_json(
     assert(0 && "Error! Action Clients must set period_ms or freq_hz in json file");
   }
 
-  auto period = std::chrono::microseconds(static_cast<int>(period_ms * 1000));
+  auto period = std::chrono::microseconds(static_cast<int64_t>(period_ms * 1000));
 
   rclcpp::QoS custom_qos_profile = get_qos_from_json(action_client_json);
